combine stories
